+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   - `binance` to get binance market data
   - `sportsdataio` to get sports data from sportsdata.io
   - `vesper` to get TVL from Vesper
+  - `linear-finance` to calculate indexes for Linear Finance
 
 ### Changed
 - Conflux adapter sending transaction optimization